pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"
[project]
name = "disha"
version = "0.0.1"
authors = [
    { name = "Mahir Patel", email = "mahirp@bu.edu" },
]
description = "A toolkit to analyze interactions among different transcription factors."
requires-python = ">=3.10"
classifiers = [
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

dependencies = [
"imutils==0.5.4",
"matplotlib==3.7.1",
"numpy==1.23.5",
"pandas==2.0.1",
"Pillow==9.5.0",
"segmentation_models==1.0.1",
"tensorflow==2.9.1",
"opencv-python-headless",
"openpyxl"]
[project.urls]
repository = "https://github.com/mahir1010/D.I.S.H.A"

[project.scripts]
disha-analyzer = "disha.disha_phy1h:disha_analyzer"
disha-verify = "disha.verify:verify_folder"